+++ imgaug/augmentables/utils.py
@@ -1,5 +1,6 @@
 from __future__ import print_function, absolute_import, division
 import numpy as np
+import six.moves as sm
 import imgaug as ia
 
 
@@ -68,3 +69,43 @@ def project_coords(coords, from_shape, to_shape):
     coords_proj[:, 0] = (coords_proj[:, 0] / from_width) * to_width
     coords_proj[:, 1] = (coords_proj[:, 1] / from_height) * to_height
     return coords_proj
+
+
+def interpolate_point_pair(point_a, point_b, nb_steps):
+    if nb_steps < 1:
+        return []
+    x1, y1 = point_a
+    x2, y2 = point_b
+    vec = np.float32([x2 - x1, y2 - y1])
+    step_size = vec / (1 + nb_steps)
+    return [(x1 + (i + 1) * step_size[0], y1 + (i + 1) * step_size[1]) for i in sm.xrange(nb_steps)]
+
+
+def interpolate_points(points, nb_steps, closed=True):
+    if len(points) <= 1:
+        return points
+    if closed:
+        points = list(points) + [points[0]]
+    points_interp = []
+    for point_a, point_b in zip(points[:-1], points[1:]):
+        points_interp.extend([point_a] + interpolate_point_pair(point_a, point_b, nb_steps))
+    if not closed:
+        points_interp.append(points[-1])
+    # close does not have to be reverted here, as last point is not included in the extend()
+    return points_interp
+
+
+def interpolate_points_by_max_distance(points, max_distance, closed=True):
+    ia.do_assert(max_distance > 0, "max_distance must have value greater than 0, got %.8f" % (max_distance,))
+    if len(points) <= 1:
+        return points
+    if closed:
+        points = list(points) + [points[0]]
+    points_interp = []
+    for point_a, point_b in zip(points[:-1], points[1:]):
+        dist = np.sqrt((point_a[0] - point_b[0]) ** 2 + (point_a[1] - point_b[1]) ** 2)
+        nb_steps = int((dist / max_distance) - 1)
+        points_interp.extend([point_a] + interpolate_point_pair(point_a, point_b, nb_steps))
+    if not closed:
+        points_interp.append(points[-1])
+    return points_interp